fix: move some things around
