+++ src/cli/lsp/proto/CMakeLists.txt
@@ -18,7 +18,7 @@ find_package(protobuf CONFIG REQUIRED)
 find_package(gRPC CONFIG REQUIRED)
 find_package(Threads)
 set(PROTO_FILES lsp.proto)
-add_library(aaltitoad_proto ${PROTO_FILES} lsp_server.cpp)
+add_library(aaltitoad_proto ${PROTO_FILES})
 target_link_libraries(aaltitoad_proto PUBLIC protobuf::libprotobuf gRPC::grpc gRPC::grpc++)
 target_include_directories(aaltitoad_proto PUBLIC ${CMAKE_CURRENT_BINARY_DIR} ${spdlog_SOURCE_DIR}/include)
 get_target_property(grpc_cpp_plugin_location gRPC::grpc_cpp_plugin LOCATION)
